refactor: CouponHistoryItem으로 폴더 이름 변경
