+++ frontend/src/@components/coupon/CouponHistoryItem/index.tsx
