+++ frontend/src/@components/coupon/CouponHistoryItem/style.tsx
